Token based testing

+++ controllers/blogs.js
@@ -1,10 +1,11 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogRouter.get("/", async (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs.map((blog) => blog.toJSON()));
-  });
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+  response.json(blogs.map((blog) => blog.toJSON()));
 });
 
 blogRouter.get("/:id", async (request, response) => {
@@ -19,15 +20,27 @@ blogRouter.get("/:id", async (request, response) => {
 // posting route
 blogRouter.post("/", async (request, response) => {
   const body = request.body;
+  console.log("body", body);
+  console.log("req token", request.token);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes === undefined ? 0 : body.likes,
+    user: user._id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.json(savedBlog.toJSON());
 });
 
@@ -49,8 +62,17 @@ blogRouter.put("/:id", (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  const user = await User.findById(decodedToken.id); ///CURRENTLY LOGGED IN USER
+  //^^ logged in user.
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    console.log("not valid user");
+    response.status(401).end();
+  }
 });
 
 module.exports = blogRouter;
